fix: move  and  in , modify overlap handling
