+++ geoutils/raster/__init__.py
@@ -23,5 +23,6 @@ from geoutils.raster.geotransformations import *  # noqa
 from geoutils.raster.multiraster import *  # noqa
 from geoutils.raster.sampling import *  # noqa
 from geoutils.raster.satimg import *  # noqa
+from geoutils.raster.tiling import *  # noqa
 
 __all__ = ["RasterType", "Raster"]
