+++ geoutils/raster/sampling.py
@@ -109,99 +109,3 @@ def subsample_array(
         return unraveled_indices
     else:
         return array[unraveled_indices]
-
-
-def _get_closest_rectangle(size: int) -> tuple[int, int]:
-    """
-    Given a 1D array size, return a rectangular shape that is closest to a cube which the size fits in.
-
-    If 'size' does not have an integer root, a rectangle is returned that is slightly larger than 'size'.
-
-    :examples:
-        >>> _get_closest_rectangle(4)  # size will be 4
-        (2, 2)
-        >>> _get_closest_rectangle(9)  # size will be 9
-        (3, 3)
-        >>> _get_closest_rectangle(3)  # size will be 4; needs padding afterward.
-        (2, 2)
-        >>> _get_closest_rectangle(55) # size will be 56; needs padding afterward.
-        (7, 8)
-        >>> _get_closest_rectangle(24)  # size will be 25; needs padding afterward
-        (5, 5)
-        >>> _get_closest_rectangle(85620)  # size will be 85849; needs padding afterward
-        (293, 293)
-        >>> _get_closest_rectangle(52011)  # size will be 52212; needs padding afterward
-        (228, 229)
-    """
-    close_cube = int(np.sqrt(size))
-
-    # If size has an integer root, return the respective cube.
-    if close_cube**2 == size:
-        return (close_cube, close_cube)
-
-    # One of these rectangles/cubes will cover all cells, so return the first that does.
-    potential_rectangles = [(close_cube, close_cube + 1), (close_cube + 1, close_cube + 1)]
-
-    for rectangle in potential_rectangles:
-        if np.prod(rectangle) >= size:
-            return rectangle
-
-    raise NotImplementedError(f"Function criteria not met for rectangle of size: {size}")
-
-
-def subdivide_array(shape: tuple[int, ...], count: int) -> NDArrayNum:
-    """
-    Create indices for subdivison of an array in a number of blocks.
-
-    If 'count' is divisible by the product of 'shape', the amount of cells in each block will be equal.
-    If 'count' is not divisible, the amount of cells in each block will be very close to equal.
-
-    :param shape: The shape of a array to be subdivided.
-    :param count: The amount of subdivisions to make.
-
-    :examples:
-        >>> subdivide_array((4, 4), 4)
-        array([[0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3]])
-
-        >>> subdivide_array((6, 4), 4)
-        array([[0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3]])
-
-        >>> subdivide_array((5, 4), 3)
-        array([[0, 0, 0, 0],
-               [0, 0, 0, 0],
-               [1, 1, 2, 2],
-               [1, 1, 2, 2],
-               [1, 1, 2, 2]])
-
-    :raises ValueError: If the 'shape' size (`np.prod(shape)`) is smallern than 'count'
-                        If the shape is not a 2D shape.
-
-    :returns: An array of shape 'shape' with 'count' unique indices.
-    """
-    try:
-        import skimage.transform
-    except ImportError:
-        raise ImportError("Missing optional dependency, skimage.transform, required by this function.")
-
-    if count > np.prod(shape):
-        raise ValueError(f"Shape '{shape}' size ({np.prod(shape)}) is smaller than 'count' ({count}).")
-
-    if len(shape) != 2:
-        raise ValueError(f"Expected a 2D shape, got {len(shape)}D shape: {shape}")
-
-    # Generate a small grid of indices, with the same unique count as 'count'
-    rect = _get_closest_rectangle(count)
-    small_indices = np.pad(np.arange(count), np.prod(rect) - count, mode="edge")[: int(np.prod(rect))].reshape(rect)
-
-    # Upscale the grid to fit the output shape using nearest neighbour scaling.
-    indices = skimage.transform.resize(small_indices, shape, order=0, preserve_range=True).astype(int)
-
-    return indices.reshape(shape)
